+++ nutype_macros/src/integer/gen/mod.rs
@@ -77,7 +77,7 @@ where
                         }
                     )
                 }
-                IntegerValidator::With(custom_is_valid_fn) => {
+                IntegerValidator::Predicate(custom_is_valid_fn) => {
                     quote!(
                         if !(#custom_is_valid_fn)(&val) {
                             return Err(#error_name::Invalid);