+++ nutype_macros/src/integer/parse.rs
@@ -73,11 +73,11 @@ where
                 item: IntegerValidator::Max(number),
                 span,
             })
-        } else if ident == "with" {
+        } else if ident == "predicate" {
             let _eq: Token![=] = input.parse()?;
             let (typed_custom_function, span) = parse_typed_custom_function::<&T>(input)?;
             Ok(SpannedIntegerValidator {
-                item: IntegerValidator::With(typed_custom_function),
+                item: IntegerValidator::Predicate(typed_custom_function),
                 span,
             })
         } else {