+++ nutype_macros/src/string/parse.rs
@@ -100,11 +100,11 @@ impl Parse for SpannedStringValidator {
                 item: StringValidator::NotEmpty,
                 span: ident.span(),
             })
-        } else if ident == "with" {
+        } else if ident == "predicate" {
             let _eq: Token![=] = input.parse()?;
             let (typed_custom_function, span) = parse_typed_custom_function_raw(input, "&str")?;
             Ok(SpannedStringValidator {
-                item: StringValidator::With(typed_custom_function),
+                item: StringValidator::Predicate(typed_custom_function),
                 span,
             })
         } else if ident == "regex" {