+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * [BREAKING] Removal of asterisk derive
 * [BREAKING] Use commas to separate high level attributes
 * [BREAKING] Traits are derived with `#[nutype(derive(Debug))]`. The regular `#[derive(Debug)]` syntax is not supported anymore.
+* [BREAKING] Rename `with = ` validator to `predicate =`
 
 ### v0.3.1 - 2023-06-30
 * Support deriving of `Deref`